CurrencyVision/requirements.txt
ultralytics
opencv-python
numpy
pyttsx3
pyyaml
num2words
